+++ cypress/integration/watchUseFieldArrayNested.tsx
@@ -3,7 +3,7 @@ describe('watchUseFieldArrayNested', () => {
     cy.visit('http://localhost:3000/watchUseFieldArrayNested');
 
     cy.get('#result').contains(
-      '[{"firstName":"Bill","keyValue":[{"name":"1a"},{"name":"1c"}]}]',
+      '[{"firstName":"Bill","keyValue":[{"name":"1a"},{"name":"1c"}],"lastName":"Luo"}]',
     );
 
     cy.get(`#nest-append-0`).click();
@@ -13,7 +13,7 @@ describe('watchUseFieldArrayNested', () => {
     cy.get(`#nest-move-0`).click();
 
     cy.get('#result').contains(
-      '[{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"1a"},{"name":"1c"},{"name":"append"}]}]',
+      '[{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"1a"},{"name":"1c"},{"name":"append"}],"lastName":"Luo"}]',
     );
 
     cy.get(`#nest-remove-0`).click();
@@ -21,7 +21,7 @@ describe('watchUseFieldArrayNested', () => {
     cy.get('#submit').click();
 
     cy.get('#result').contains(
-      '[{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}]}]',
+      '[{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}],"lastName":"Luo"}]',
     );
 
     cy.get('#prepend').click();
@@ -30,7 +30,7 @@ describe('watchUseFieldArrayNested', () => {
     cy.get('#insert').click();
 
     cy.get('#result').contains(
-      '[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}]}]',
+      '[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}],"lastName":"Luo"}]',
     );
 
     cy.get(`#nest-append-0`).click();
@@ -40,14 +40,14 @@ describe('watchUseFieldArrayNested', () => {
     cy.get(`#nest-move-0`).click();
 
     cy.get('#result').contains(
-      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"append"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}]}]',
+      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"append"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}],"lastName":"Luo"}]',
     );
 
     cy.get('#nest-remove-3').click();
     cy.get('#nest-remove-3').click();
 
     cy.get('#result').contains(
-      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"append"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"append"}]}]',
+      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"append"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"append"}],"lastName":"Luo"}]',
     );
 
     cy.get('#nest-remove-all-3').click();
@@ -56,7 +56,7 @@ describe('watchUseFieldArrayNested', () => {
     cy.get('#nest-remove-all-0').click();
 
     cy.get('#result').contains(
-      '[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[]}]',
+      '[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[],"lastName":"Luo"}]',
     );
 
     cy.get('#remove').click();