+++ src/__tests__/useFieldArray.test.tsx
@@ -638,7 +638,7 @@ describe('useFieldArray', () => {
       act(() => {
         result.current.reset();
       });
-      expect(result.current.fields).toEqual([{ id: '3', value: 'default' }]);
+      expect(result.current.fields).toEqual([{ id: '4', value: 'default' }]);
     });
 
     it('should reset with field array with shouldUnregister set to false', () => {
@@ -665,7 +665,7 @@ describe('useFieldArray', () => {
         result.current.reset();
       });
 
-      expect(result.current.fields).toEqual([{ id: '3', value: 'default' }]);
+      expect(result.current.fields).toEqual([{ id: '4', value: 'default' }]);
 
       act(() => {
         result.current.reset({
@@ -673,7 +673,7 @@ describe('useFieldArray', () => {
         });
       });
 
-      expect(result.current.fields).toEqual([{ id: '5', value: 'data' }]);
+      expect(result.current.fields).toEqual([{ id: '6', value: 'data' }]);
     });
 
     it('should reset with async', async () => {