+++ src/useForm.ts
@@ -25,6 +25,8 @@ import isFunction from './utils/isFunction';
 import isHTMLElement from './utils/isHTMLElement';
 import isMultipleSelect from './utils/isMultipleSelect';
 import isNullOrUndefined from './utils/isNullOrUndefined';
+import isObject from './utils/isObject';
+import isPrimitive from './utils/isPrimitive';
 import isProxyEnabled from './utils/isProxyEnabled';
 import isRadioInput from './utils/isRadioInput';
 import isRadioOrCheckboxFunction from './utils/isRadioOrCheckbox';
@@ -109,7 +111,7 @@ export function useForm<
     new Subject<{
       name?: InternalFieldName;
       type?: EventType;
-      value?: unknown;
+      formValues: unknown;
     }>(),
   );
   const controllerSubjectRef = React.useRef(
@@ -389,10 +391,7 @@ export function useForm<
       currentNames: FieldName<TFieldValues>[] = [],
     ) => {
       const { errors } = await resolverRef.current!(
-        getFieldsValues(
-          fieldsRef,
-          shouldUnregister ? {} : defaultValuesRef.current,
-        ),
+        getFieldsValues(fieldsRef),
         contextRef.current,
         {
           criteriaMode,
@@ -616,7 +615,7 @@ export function useForm<
       : setFieldValue(name, value, options, true, !field);
 
     isFieldWatched(name) && formStateSubjectRef.current.next({});
-    watchSubjectRef.current.next({ name, value });
+    watchSubjectRef.current.next({ name, formValues: getValues() });
   };
 
   const handleChange: ChangeHandler = React.useCallback(
@@ -670,7 +669,7 @@ export function useForm<
             watchSubjectRef.current.next({
               name,
               type,
-              value: inputValue,
+              formValues: getValues(),
             });
           return (
             shouldRender &&
@@ -686,10 +685,7 @@ export function useForm<
 
         if (resolverRef.current) {
           const { errors } = await resolverRef.current(
-            getFieldsValues(
-              fieldsRef,
-              shouldUnregister ? {} : defaultValuesRef.current,
-            ),
+            getFieldsValues(fieldsRef),
             contextRef.current,
             {
               criteriaMode,
@@ -726,7 +722,7 @@ export function useForm<
           watchSubjectRef.current.next({
             name,
             type,
-            value: inputValue,
+            formValues: getValues(),
           });
         shouldRenderBaseOnError(
           name,
@@ -746,12 +742,10 @@ export function useForm<
       | FieldPath<TFieldValues>
       | ReadonlyArray<FieldPath<TFieldValues>>,
   ) => {
-    const values = isMountedRef.current
-      ? getFieldsValues(
-          fieldsRef,
-          shouldUnregister ? {} : defaultValuesRef.current,
-        )
-      : defaultValuesRef.current;
+    const values = {
+      ...defaultValuesRef.current,
+      ...getFieldsValues(fieldsRef),
+    };
 
     return isUndefined(fieldNames)
       ? values
@@ -767,10 +761,7 @@ export function useForm<
       if (resolver) {
         const { errors } = await resolverRef.current!(
           {
-            ...getFieldsValues(
-              fieldsRef,
-              shouldUnregister ? {} : defaultValuesRef.current,
-            ),
+            ...getFieldsValues(fieldsRef),
             ...values,
           },
           contextRef.current,
@@ -824,15 +815,20 @@ export function useForm<
   };
 
   const watchInternal: WatchInternal<TFieldValues> = React.useCallback(
-    (fieldNames, defaultValue, isGlobal) => {
+    (fieldNames, defaultValue, isGlobal, formValues) => {
       const isArrayNames = Array.isArray(fieldNames);
-      const fieldValues = isMountedRef.current
-        ? getFieldsValues(fieldsRef, defaultValuesRef.current)
+      const fieldValues = formValues
+        ? formValues
+        : isMountedRef.current
+        ? {
+            ...defaultValuesRef.current,
+            ...getFieldsValues(fieldsRef),
+          }
         : isUndefined(defaultValue)
         ? defaultValuesRef.current
         : isArrayNames
-        ? defaultValue || {}
-        : { [fieldNames as string]: defaultValue };
+        ? defaultValue
+        : { [fieldNames as InternalFieldName]: defaultValue };
 
       if (isUndefined(fieldNames)) {
         isGlobal && (isWatchAllRef.current = true);
@@ -841,9 +837,9 @@ export function useForm<
 
       const result = [];
 
-      for (const fieldName of isArrayNames ? fieldNames : [fieldNames]) {
-        isGlobal && watchFieldsRef.current.add(fieldName as string);
-        result.push(get(fieldValues, fieldName as string));
+      for (const fieldName of convertToArrayPayload(fieldNames)) {
+        isGlobal && watchFieldsRef.current.add(fieldName as InternalFieldName);
+        result.push(get(fieldValues, fieldName as InternalFieldName));
       }
 
       return isArrayNames ? result : result[0];
@@ -899,6 +895,7 @@ export function useForm<
 
         watchSubjectRef.current.next({
           name: inputName,
+          formValues: getValues(),
         });
       }
     }
@@ -1023,10 +1020,7 @@ export function useForm<
         e.persist && e.persist();
       }
       let hasNoPromiseError = true;
-      let fieldValues = getFieldsValues(
-        fieldsRef,
-        shouldUnregister ? {} : defaultValuesRef.current,
-      );
+      let fieldValues = getFieldsValues(fieldsRef);
 
       formStateSubjectRef.current.next({
         isSubmitting: true,
@@ -1127,6 +1121,32 @@ export function useForm<
     [],
   );
 
+  const registerAbsentFields = <T extends DefaultValues<TFieldValues>>(
+    value: T,
+    name = '',
+  ): void => {
+    !get(fieldsRef.current, name) &&
+      (isPrimitive(value) ||
+        (isWeb && (value instanceof File || value instanceof Date))) &&
+      set(fieldsRef.current, name, {
+        _f: {
+          ref: { name, value },
+          value,
+          name,
+        },
+      });
+
+    if (Array.isArray(value) || isObject(value)) {
+      if (name && !get(fieldsRef.current, name)) {
+        set(fieldsRef.current, name, Array.isArray(value) ? [] : {});
+      }
+
+      for (const key in value) {
+        registerAbsentFields(value[key], name + (name ? '.' : '') + key);
+      }
+    }
+  };
+
   const reset: UseFormReset<TFieldValues> = (values, keepStateOptions = {}) => {
     const updatedValues = values || defaultValuesRef.current;
 
@@ -1138,12 +1158,10 @@ export function useForm<
             ? field._f.refs[0]
             : field._f.ref;
 
-          if (isHTMLElement(inputRef)) {
-            try {
-              inputRef.closest('form')!.reset();
-              break;
-            } catch {}
-          }
+          try {
+            isHTMLElement(inputRef) && inputRef.closest('form')!.reset();
+            break;
+          } catch {}
         }
       }
     }
@@ -1159,7 +1177,7 @@ export function useForm<
       });
 
       watchSubjectRef.current.next({
-        value: { ...updatedValues },
+        formValues: { ...updatedValues },
       });
 
       fieldArraySubjectRef.current.next({
@@ -1168,14 +1186,17 @@ export function useForm<
       });
     }
 
+    !keepStateOptions.keepDefaultValues &&
+      registerAbsentFields({ ...updatedValues });
+
     resetFromState(keepStateOptions, values);
-    isMountedRef.current = false;
   };
 
   const setFocus: UseFormSetFocus<TFieldValues> = (name) =>
     get(fieldsRef.current, name)._f.ref.focus();
 
   React.useEffect(() => {
+    registerAbsentFields(defaultValuesRef.current);
     const formStateSubscription = formStateSubjectRef.current.subscribe({
       next(formState) {
         if (shouldRenderFormState(formState, readFormStateRef.current, true)) {
@@ -1212,7 +1233,8 @@ export function useForm<
       !isHTMLElement(ref) || !document.contains(ref);
 
     isMountedRef.current = true;
-    unregisterFieldsNamesRef.current.forEach((name) => {
+
+    for (const name of unregisterFieldsNamesRef.current) {
       const field = get(fieldsRef.current, name) as Field;
 
       field &&
@@ -1220,7 +1242,8 @@ export function useForm<
           ? field._f.refs.every(isLiveInDom)
           : isLiveInDom(field._f.ref)) &&
         unregister(name as FieldPath<TFieldValues>);
-    });
+    }
+
     unregisterFieldsNamesRef.current = new Set();
   });
 
@@ -1246,7 +1269,7 @@ export function useForm<
         defaultValuesRef,
         fieldArrayDefaultValuesRef,
         unregister,
-        shouldUnmountUnregister: shouldUnregister,
+        shouldUnmount: shouldUnregister,
       }),
       [],
     ),